+++ scripts/plp2gtopt/manli_parser.py
@@ -52,7 +52,7 @@ class ManliParser(BaseParser):
                     continue
 
                 # Initialize numpy arrays
-                blocks = np.empty(num_blocks, dtype=np.int16)
+                blocks = np.empty(num_blocks, dtype=np.int32)
                 p_max_ab = np.empty(num_blocks, dtype=np.float64)
                 p_max_ba = np.empty(num_blocks, dtype=np.float64)
                 operational = np.empty(num_blocks, dtype=np.int8)
@@ -67,7 +67,7 @@ class ManliParser(BaseParser):
                     blocks[i] = self._parse_int(parts[0])  # Block number
                     p_max_ab[i] = self._parse_float(parts[1])  # Max flow AB
                     p_max_ba[i] = self._parse_float(parts[2])  # Max flow BA
-                    operational[i] = np.int16(parts[3] == "T")  # Operational status
+                    operational[i] = np.int32(parts[3] == "T")  # Operational status
 
                 # Store complete data
                 manli = {
